Add Fibonacci lattice functions, move maths functions into new file
